backup: backup-files.sh -> backup-all-files.sh
I'm semi-retiring the S3{,QL}-based setup which involves backing up the
bulk of the data file-wise. The new plan is to backup the bulk of the
data using borg onto a remote-shell-style service (LimaLabs or
rsync.net) except for local borg repos on the tank which will be
mirrored using rsync.