isolate: isolate.sh: provision and/or clear zram disks on each VM start
If a block-device-backed disk points to a block device with path of
"/dev/zram[0-9]" or "/dev/zram/...", attempt to create and/or clear a
matching zram device. In the first form, the device must already exist
(it is impossible to create a zram device with a predefined number).
In the second form, a non-existing device may be created (see below) and
symlinked to the given path.

If the target block device name starts or ends with a size specification
(e. g. "/dev/zram/win10-cache-20G"), a non-existing zram device will be
created with given size. If the VM name starts with "scratch-", an
existing zram device will be cleared. Otherwise, no action is taken.