system: udev-container.sh: rules: check for non-empty DEVNODE
Ignore devices that do not have a device node file in /dev regardless of
any future checks (which may be overly lax, such as `SUBSYSTEM=="drm"`).

If a device does not have a /dev node, 1) the entire script is pointless
and 2) it will break and complain loudly while parsing arguments.