system: udev-container.sh: refine target readiness check
Check for either `starting` or `running` text returns instead of
checking for a non-zero exit code. With the way we rely on readiness
notifications from systemd-nspawn (which are only usable if
`--notify-ready=` or `NotifyReady=` is set), these notifications are
only emitted after reaching the final target (which may be a while).

There is no need to wait that long; accept actions as soon as the bus is
available even if the final target has not yet been reached. This way we
can hopefully catch the massive flurry of the change events that arrive
during container bootup (for reasons that I do not understand).

This will not move the staged action execute script any earlier, so all
those events would be replayed twice, but this is better than the
opposite.