misc: systemd: tailscaled.service: add After= overrides for common network managers
We order tailscaled.service `Before=network.target` to fit semantics
of Tailscale being used as load-bearing network provider (sometimes...),
but this leads to tailscaled hanging on shutdown because it expects to
be able to talk to the Internet on shutdown (and the software providing
the actual connectivity usually dies first).

There is no clean way to do this like a `network-real-actually.target`,
so instead order tailscaled after known network managers to implement
this partial ordering.